+++ src/main/java/com/ciphertool/engine/fitness/cipherkey/FrequencyCorpusFitnessEvaluator.java
@@ -16,6 +16,7 @@
 
 package com.ciphertool.engine.fitness.cipherkey;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -249,7 +250,7 @@ public class FrequencyCorpusFitnessEvaluator implements FitnessEvaluator {
 	}
 
 	@Override
-	public Double evaluate(Chromosome chromosome) {
+	public BigDecimal evaluate(Chromosome chromosome) {
 		String currentSolutionString = WordGraphUtils.getSolutionAsString((CipherKeyChromosome) chromosome).substring(0, lastRowBegin);
 
 		Map<Integer, Match> matchMap = new HashMap<Integer, Match>(currentSolutionString.length());
@@ -367,7 +368,7 @@ public class FrequencyCorpusFitnessEvaluator implements FitnessEvaluator {
 			}
 		}
 
-		return Double.valueOf(fitness);
+		return BigDecimal.valueOf(fitness);
 	}
 
 	@Override
