+++ src/main/java/com/ciphertool/engine/fitness/cipherkey/LetterAndWordMarkovFitnessEvaluator.java
@@ -19,6 +19,7 @@
 
 package com.ciphertool.engine.fitness.cipherkey;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -296,13 +297,13 @@ public class LetterAndWordMarkovFitnessEvaluator implements FitnessEvaluator {
 	}
 
 	@Override
-	public Double evaluate(Chromosome chromosome) {
+	public BigDecimal evaluate(Chromosome chromosome) {
 		double total = 0.0;
 		total += (frequencyWeight == 0.0) ? 0.0 : (frequencyWeight * evaluateFrequency(chromosome));
 		total += (letterNGramWeight == 0.0) ? 0.0 : (letterNGramWeight * evaluateMarkovModel(chromosome));
 		total += (wordNGramWeight == 0.0) ? 0.0 : (wordNGramWeight * evaluateNGram(chromosome));
 
-		return total;
+		return BigDecimal.valueOf(total);
 	}
 
 	public Double evaluateFrequency(Chromosome chromosome) {
